requirements.txt
Django>=4.2.5
whitenoise>=6.5.0
django-ckeditor
yfinance
